requirements.txt
alembic==1.5.4
attrs==21.2.0
autopep8==1.5.5
blinker==1.5
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.0.1
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
coverage==7.0.3
cycler==0.11.0
decorator==5.1.1
Fiona==1.8.22
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
fonttools==4.38.0
future==0.18.3
GeoAlchemy2==0.13.1
geopandas==0.12.2
greenlet==2.0.1
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.4.4
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.6.3
munch==2.5.0
numpy==1.24.1
packaging==20.9
pandas==1.5.3
Pillow==9.4.0
pluggy==0.13.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.6.0
pyparsing==2.4.7
pyproj==3.4.1
pytest==7.1.1
pytest-cov==4.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2022.7.1
ratelim==0.1.6
requests==2.25.1
Rtree==1.0.1
shapely==2.0.0
six==1.15.0
SQLAlchemy==1.4.16
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.4
Werkzeug==1.0.1
